middleware/requirements.txt
Flask>=3.0,<4.0
PyGithub>=2.3,<3.0
jsonschema>=4.23,<5.0
